add a new config named "completeIssuer".
This new config handles the command completion when the tab key is
pressed. It differs from `'completeHandle'`. `'completeIssuer'` will
just trigger the calculation for completion, and the result is returned
asynchronously, after which the controller's
`'showCompletion(promptText, completions)'` can be invoked with the
result. `'completeHandle'` will retrieve the result synchronously, and
show the result. If `'completeHandle'` exists, `'completeIssuer'` is
ignored. A typical usage of `'completeIssuer'` is that the completion is
retrieved from the server using ajax or WebSocket asynchronously.