ability to modify previous messages
added in the ability to access previous messages in the console.
For example
container.history.get(0); returns an instance of MessageHandle for the most recent message in the console.
I added a setTest(str) method to MessageHandle so  you can change the text in that particular mesg. The message handle will continue to point to the message regardless of now many new messages are added. I needed this so i could make an interactive ..... which adds a few dots per second as my client waits for the server to reply.

Also connected the history into clearScreen so it is cleared accordingly.